--- a/memview/memview.go
+++ b/memview/memview.go
@@ -150,11 +150,20 @@ func (mv MemView) Index(start int64, sep []byte) int64 {
 
 	// Iteratively search for the target, keeping in mind that the target may be
 	// spread over multiple slices in mv.buf.
+	//
+	// TODO: this only works correctly for search strings that do not have a repeated
+	// prefix. To work correctly, we would have to back up to the point at which
+	// the needle *could* have started after an incomplete match.
+	//
+	// However, we only use this method to search for strings without a repeated prefix:
+	// GET, POST, DELETE, HEAD, PUT, PATCH, CONNECT, OPTIONS, TRACE, HTTP/1.1 and HTTP/1.0
 	needle := sep
 	needleIndex := 0
 	for b := startBuf; b < len(mv.buf); b++ {
 		haystack := mv.buf[b]
-		for i := startOffset; i < len(haystack); i++ {
+		// Check remainder of needle if overlap from last buffer
+		var i int = 0
+		for i = startOffset; i < len(haystack) && needleIndex > 0; i++ {
 			if haystack[i] == needle[needleIndex] {
 				needleIndex += 1
 				if needleIndex == len(needle) {
@@ -168,6 +177,30 @@ func (mv MemView) Index(start int64, sep []byte) int64 {
 				needleIndex = 0
 			}
 		}
+
+		// Did we reach the end of the buffer already?
+		if i < len(haystack) {
+			// If not, efficient check of remaining portion of haystack
+			found := bytes.Index(haystack[i:], needle)
+			if found != -1 {
+				return currIndex + int64(found)
+			}
+
+			// Check the end of the haystack for the start of the needle
+			// (but not the whole thing, or we would have found it in the call above.)
+			needleStart := len(haystack) - len(needle) + 1
+			if i < needleStart {
+				i = needleStart
+			}
+			for ; i < len(haystack); i++ {
+				if haystack[i] == needle[needleIndex] {
+					needleIndex += 1
+				} else {
+					needleIndex = 0
+				}
+			}
+		}
+
 		// Searched all of buffer
 		currIndex += int64(len(haystack) - startOffset)
 		startOffset = 0