Added benchmark tests.
Added a failing test not relevant to our actual usage.

Old implementation:
```
BenchmarkIndexSmall-8   	   94458	     59803 ns/op
BenchmarkIndexLarge-8   	     337	  17548862 ns/op
```

New implementation:
```
BenchmarkIndexSmall-8   	  435463	     12523 ns/op
BenchmarkIndexLarge-8   	    1570	   3629018 ns/op
```

This could be improved further by searching in parallel for a set of strings; the benchmark roughly reflects how we actually use it.