@mgritter 2021-07-31
We are generally in a situation where our buffers are large and our search strings are small.  So it's probably worth using bytes.Index() which has a platform-optimized implementation, and then this checking the len(needle) wraparound cases at each boundary.

That would also solve the correctness problem.

However, it's probably still less efficient than parallel search for the set of strings we're interested in.