@mgritter 2021-08-02
New benchmarking:

Previous version in PR:
```
BenchmarkIndexSmall-8   	 3002158	     11870 ns/op
BenchmarkIndexLarge-8   	    9596	   3624710 ns/op
```

Most recent:
```
BenchmarkIndexSmall-8   	12983755	      2619 ns/op
BenchmarkIndexLarge-8   	   23590	   1597876 ns/op
```

This is by using `bytes.Index()` instead of rolling our own.

I tried a version of parallel search using `regexp` but it was much slower (about 10000/30 seconds instead of 23590) so that is not included in the PR.