@mgritter 2021-08-02
Re-requested from @thatplguy due to significant changes.